Separate tests + complete rating
1. The tests are now separated into individual drivers and individual tests
2. Add test for verifying whether the box is selected.
3. Picked up the locator specifically related to the username or evaluation creation time in html.
